Works till'clipping migration dataset to extent'

+++ copro/utils.py
@@ -25,18 +25,21 @@ def get_geodataframe(config, root_dir, geometry='geometry', crs='WGS84'):
 
     gdf.rename(columns = {'M2001':'2001', 'M2002':'2002', 'M2003':'2003', 'M2004':'2004', 
                           'M2005':'2005', 'M2006':'2006', 'M2007':'2007', 'M2008':'2008', 
-                          'M2009':'2009', 'M2010':'2010', 'M2011':'2011', 'M2012':'2012', 
-                          'M2013':'2013', 'M2014':'2014', 'M2015':'2015'}, inplace = True)
+                          'M2009':'2009', 'M2010':'2010', 'M2011':'2011', 'M2012':'2012', 'M2013':'2013', 'M2014':'2014', 'M2015':'2015'}, inplace = True)
+
 
     # Reorganise GDF so that years can be selected 
 
-    gdf = gdf.melt(id_vars=['GID_2', 'geometry'], value_vars=['2001', '2002', '2003', '2004', '2005', 
-                                                              '2006', '2007', '2008', '2009', '2010', 
-                                                              '2011', '2012', '2013', '2014', '2015'], 
-                                                              var_name='year', value_name='net_migration')
+    gdf = gdf.melt(id_vars=['GID_2', 'geometry'], value_vars= ['2001', '2002', '2003', '2004', '2005', 
+                                                              '2006', '2007', '2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015'], var_name='year', value_name='net_migration')
 
-# TESTED print(gdf) --> works
+    
+    gdf.year = gdf.year.astype(int)
+
+    # TESTED print(gdf) --> works
     return gdf
+
+    
 
 def show_versions():
     """click.echos the version numbers by the main python-packages used.
